# docs/APISpec.json
{
  "swagger": "2.0",
  "info": {
    "title": "APISpec",
    "version": "v0.0.0"
  },
  "host": "localhost:8080",
  "schemes": [
    "http"
  ],
  "basePath": "/",
  "paths": {
    "/": {
      "get": {
        "summary": "Ping Server",
        "description": "Pings the main server endpoint to confirm that the server is actually running.",
        "operationId": "35d8fc2e-ddf9-4e33-af98-a5df1b94bc9b",
        "parameters": [
          {
            "type": "string",
            "default": "",
            "name": "Cookie",
            "required": false,
            "in": "header"
          }
        ],
        "responses": {
          "default": {
            "description": "no response description was provided for this operation"
          }
        }
      }
    },
    "/accounts/": {
      "get": {
        "summary": "Ping Accounts",
        "description": "Pings the main server endpoint to confirm that the server is actually running.",
        "operationId": "47250fa6-ca3b-424d-bbdd-bb9b708aa8d8",
        "parameters": [
          {
            "type": "string",
            "default": "",
            "name": "Cookie",
            "required": false,
            "in": "header"
          }
        ],
        "responses": {
          "default": {
            "description": "no response description was provided for this operation"
          }
        },
        "security": [
          {
            "oauth_2_auth": []
          }
        ]
      },
      "delete": {
        "summary": "Delete Account",
        "description": "",
        "operationId": "467e7891-8571-42e4-965d-8d4861e1a88f",
        "parameters": [
          {
            "type": "string",
            "default": "",
            "name": "Cookie",
            "required": false,
            "in": "header"
          }
        ],
        "responses": {
          "default": {
            "description": "no response description was provided for this operation"
          }
        },
        "security": [
          {
            "oauth_2_auth": []
          }
        ]
      }
    },
    "/accounts/auth/register": {
      "post": {
        "summary": "Register Account",
        "description": "",
        "operationId": "f2a6cf7f-4eca-4573-a308-084ff63aab1c",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "string",
            "default": "",
            "name": "Cookie",
            "required": false,
            "in": "header"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "default": "{\"type\":\"INDIVIDUAL\",\"username\":\"test_username\",\"password\":\"test_password\"}"
            },
            "in": "body",
            "name": "body"
          }
        ],
        "responses": {
          "default": {
            "description": "no response description was provided for this operation"
          }
        }
      }
    },
    "/accounts/auth/login": {
      "post": {
        "summary": "Login Account",
        "description": "",
        "operationId": "015c5f2f-b32a-4dca-8648-f945d7ac9c48",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "string",
            "default": "",
            "name": "Cookie",
            "required": false,
            "in": "header"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "default": "{\"type\":\"INDIVIDUAL\",\"username\":\"test_username\",\"password\":\"test_password\"}"
            },
            "in": "body",
            "name": "body"
          }
        ],
        "responses": {
          "default": {
            "description": "no response description was provided for this operation"
          }
        }
      }
    },
    "/accounts/type": {
      "put": {
        "summary": "Change Account Type",
        "description": "",
        "operationId": "562b772f-2a3a-4b84-ac8d-1bb7e89e3e81",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "string",
            "default": "",
            "name": "Cookie",
            "required": false,
            "in": "header"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "default": "{\"type\":\"INDIVIDUAL\"}"
            },
            "in": "body",
            "name": "body"
          }
        ],
        "responses": {
          "default": {
            "description": "no response description was provided for this operation"
          }
        },
        "security": [
          {
            "oauth_2_auth": []
          }
        ]
      }
    },
    "/accounts/password": {
      "put": {
        "summary": "Reset Account Password",
        "description": "",
        "operationId": "724075f0-4432-4fd4-9499-d480b955dd2e",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "string",
            "default": "",
            "name": "Cookie",
            "required": false,
            "in": "header"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "default": "{\"type\":\"INDIVIDUAL\",\"username\":\"test_username\",\"password\":\"test_password\"}"
            },
            "in": "body",
            "name": "body"
          }
        ],
        "responses": {
          "default": {
            "description": "no response description was provided for this operation"
          }
        },
        "security": [
          {
            "oauth_2_auth": []
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "oauth_2_auth": {
      "type": "oauth2",
      "scopes": {}
    }
  },
  "tags": []
}